vlq cleanup
